session_log: carve history.rs (pure move)
Read-back of conversation history from a session log directory:
ConversationTurn + recent_conversation (voice_log/user_transcript turns
for presence context), search_voice_entries keyword search, and the
recent_entries raw JSONL tail. The recent_* and search_* tests ride
along.
